Move code from wq.timeutil to util.time
